memory-mcp-server/requirements.txt
mcp
qdrant-client
fastembed
python-dotenv
